Claude/requirements.txt
anthropic>=0.7.0
python-dotenv>=1.0.0